Extract Launchplane preview admin CLI
